+++ tests/files/src/0522_complex_cond_loop.php
@@ -1,0 +1,7 @@
+<?php
+call_user_func(function() {
+    $dir = opendir('.');
+    while ($dir && (false !== ($file = readdir($dir)))) {
+        echo $file;
+    }
+});