+++ tests/misc/ast/src/021_return.php
@@ -1,5 +1,6 @@
 <?php
 // Phan's BlockAnalysisVisitor is already aware that the branch with `return $x` is unreachable.
-function f344(?string $x) : int {
+/** @param ?string $x */
+function f344($x) : int {
     if (is_string($x)) { } else { return $x; } return 2;
 }