+++ tests/files/src/0524_closure_ambiguous.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Bar524 {
+    class Closure {}
+}
+
+namespace Foo524 {
+
+    use Closure;
+    /** @param Closure():void $closure */
+    function test524a($closure) {}
+    /** @param \Closure():void $closure */
+    function test524b($closure) {}
+}
+
+namespace Baz524 {
+    /** @param Closure():void $closure */
+    function test524a($closure) {}
+    /** @param \Closure():void $closure */
+    function test524b($closure) {}
+}
+
+namespace {
+
+    use Bar524\Closure;
+    /** @param Closure():void $closure */
+    function test524a($closure) {}
+    /** @param \Closure():void $closure */
+    function test524b($closure) {}
+}