+++ NEWS.md
@@ -14,6 +14,8 @@ New features(Analysis)
 
   Named subpatterns, non-capturing patterns, and regular expression options are not supported yet.
   Phan will just infer a more generic type such as `string[]` (depending on the bit flags).
++ Warn about ambiguous uses of `Closure():void` in phpdoc.
+  Also, make that syntax count as a reference to `use Closure;` in that namespace.
 
 Bug fixes
 + Fix a crash seen when using a temporary expression in a write context. (#1915)